dev-requirements.txt
pytest>=2.7.3
pytest-capturelog
pytest-flask
pylint
tox
zest.releaser[recommended]
-e .
